added bouncing square (simplified dice)

+++ include/init.h
@@ -1,8 +1,8 @@
 #ifndef __INIT_H__
 #define __INIT_H__
 
-extern const int screen_width;
-extern const int screen_height;
+extern const float screen_width;
+extern const float screen_height;
 
 extern void init();
